+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased
+
+- FIX [TBD](https://github.com/MatthewHerbst/react-to-print/pull/TBD): When `copyShadowRoots` is enabled, also copy constructable stylesheets (`adoptedStyleSheets`) into the print output. Previously styles applied this way (common with web component frameworks such as Lit) were missing ([770](https://github.com/MatthewHerbst/react-to-print/issues/770))
+
 ## 3.3.0 (February 22nd, 2026)
 
 - CHORE [838](https://github.com/MatthewHerbst/react-to-print/pull/838): Update most dev dependencies